test(app): isolate dynamic slot coverage (#28102)
